requirements.txt
click==8.1.3
Flask==2.2.3
greenlet==2.0.2
importlib-metadata==6.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
PyMySQL==1.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.9
tomli==2.0.1
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
